feat(evaluate): divide realms into core and contrib
BREAKING CHANGE: only json and map-set realms and still part of the package.
apidom, minim and immutable are no longer part of npm distributions but live in
contrib directory in the GitHub repo.
