+++ src/evaluate/index.js
@@ -171,7 +171,12 @@ const evaluate = (
       reason: error.message,
     });
 
-    if (error instanceof JSONPointerEvaluateError) {
+    // Check both instanceof and error name to support errors from different package instances
+    // (e.g., contrib realms importing from built package while evaluate runs from source)
+    if (
+      error instanceof JSONPointerEvaluateError ||
+      (error?.name && /^JSONPointer(Type|Key|Index|Evaluate)Error$/.test(error.name))
+    ) {
       throw error;
     }
