--- src/evaluate/realms/apidom/index.js
@@ -1,2 +1,0 @@
-export { default } from './realm.js';
-export { default as evaluate } from './evaluate.js';
