+++ test/loader.js
@@ -1,0 +1,4 @@
+import { register } from 'node:module';
+import { pathToFileURL } from 'node:url';
+
+register('./loader-hooks.js', pathToFileURL('./test/'));
